Added support and documentation for the required --site-dir=DIR command line option.
Note that this option implements a slight departure from the original
'make_site.sh' program in that it requires the user to explicitly name the
directory in which the site is created. The 'make_site.sh' program had a
$WEBSITE_ROOT config param under which the site directory was always created
in the following form:

    $WEBSITE_ROOT/$DOMAIN_NAME

While that is certainly a reasonable way of managing web sites, that
approach forces the directory naming policy on the user.