Added infrastructure (but not implementation logic) for new required command line opt: --site-conf=FILE.
This option is where the user wants us to write the Apache 2.0.x virtual
host snippet for the Bedrock site being configured.

Also added infrastructure (but again, no impl logic yet) for fine-grained
"overwrite" opts that affect overwriting of the specified file:

    --overwrite-site-conf
    --nooverwrite-site-conf

Removed documentation that suggested that the default name of the site
config file would be $DOMAIN_NAME.conf. While that scheme is certainly a
sane default to use, keeping that as the default behavior of the program
would require a new required option for the apache configuration directory.

Note that this option implements a slight departure from the original
'make_site.sh' program in that it requires the user to explicitly name the
apache config snippet file. The 'make_site.sh' program had an
$APACHE2_CONFIG_DIR config param under which the site conf file was always
created in the following form:

    $APACHE2_CONFIG_DIR/$DOMAIN_NAME.conf

I've change the behavior to make the program behave correctly without having
to inflict a specific naming policy on the user, and without him having to
understand very much about the program. What the user specifies is what the
user gets.