Checkpoint commit. Implemented partial functionality for specifying a separate apache auth db include file (and added it to the "(no)overwrite" framework). Documented all currently existing options, along with some caveats.
Also beefed up overwrite options summary display to only display summary
overwrite data that applies. Until now, all overwrite options applied. But
now we have an apache auth db include file option that is strictly optional,
so we only want to provide overwrite flag summary data for this file if the
user specified it. Doing otherwise would be both confusing to the user, and
ugly in the display (since we would not have a file name to refer to).