Checkpoint commit. Not yet tested (not even sure if it will compile).
Added utily function __create_directories_or_die(). See comments in code for
details.

Fleshed out the rest of maybe_create_apache_site_config_file_or_die() with
logic to write the apache authentication directives. See the function
documentation for the full skinny. Also refactored the function to take
"named" parameters (instead of positional parameters) as there are 12(!)
parameters, only 4 of which are required.

The long list of params is due to the large number of authentication
properties we need to have values for in order to write out the apache
config file (only some which we can provide sane defaults for). We could
bundle all of the params up into a single structure or object which is
passed to the function instead, but that seems like overkill for a function
not intended for external consumption; the named params approach keeps the
caller self-documenting, and the function is self-contained, otherwise well
documented, and has much internal sanity checking so shouldn't be a problem.

This work was partly done in anticipation of externalizing each major
feature of 'bedrock-create-site' into a separate tool.

Added variable to receive authentication option params, but have not yet
provided any wiring to allow the user to specify those options.