Committing changes to allow the user to specify the owner and group that should own /most/ of the file and directories created by create-bedrock-site. These features are used via the two new command line options:
    -o, --owner=OWNER
    -g, --group=GROUP

The OWNER and GROUP arguments may be the respective numer uid and gid, the
user name and group name, or a mixture (one numeric and one
non-numeric). Postel's Prescription, baby!

The Apache site config file and the apache auth db file (and any created
parent directories) are not affected by by these options; they will /always/
be owned by the effective uid (euid) and effective group id (egid) of the
user invoking the program. This allows the program to be run as root to set
file and directory ownership to someone non-root without allowing that
non-root user to change the apache site config or apache auth db config.

Any created symlinks are not affected by the above options, but they're
symlinks so it doesn't matter one way or the other.

See man page POD for all the gory details, including the new entries in the
BUGS section.

To support the above features, we've added the following new functions:

    __poor_mans_touch()
    __chown_chgrp_all()

See comments in the code for details