Changed finalize_childen to be a regular method.
The reason for this is that we want to be able to serialize and deserialize
parsed objects.

In addition, the call to output() from t.pl was removed -- it is superfluous,
at least as far as our small test case is concerned.  In the future, we should
revisit (and/or rename) this set of methods: parse, get_parse_object, output,
finalize, and finalize_children.

Removed some of the older logging code.